pyproject.toml
[tool.black]
line-length = 120

[tool.isort]
line_length = 120
profile = "black"

[tool.poetry]
name = "second"
version = "0.1.0"
description = ""
authors = ["COCOLMAN <iam@hyunsol.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
python-dotenv = "^1.0.0"
cryptography = "^41.0.7"
sqlalchemy = "2.0.20"
pyjwt = "^2.8.0"
aiohttp = {extras = ["speedups"], version = "^3.9.1"}
aiomysql = "^0.2.0"
greenlet = "^3.0.3"

[tool.poetry.group.manage.dependencies]
alembic = "^1.13.0"

[tool.poetry.group.lambda.dependencies]
mangum = "^0.17.0"

[tool.poetry.group.dev.dependencies]
boto3 = "^1.34.11"
boto3-stubs = {extras = ["ssm"], version = "^1.34.11"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
